requirements.txt
cfbd == 4.5.2
ipykernel == 6.28.0
pillow == 10.2.0
requests == 2.31.0
pandas == 2.1.4